ADD: Aggiunto controllo nel cleaner per i ticket ancora in setup

+++ bin/HandlingFunctions.js
@@ -206,16 +206,18 @@ async function cleanerDatabase(client) {
       const channel = await guild.channels.fetch(value.channelId);
       await channel.messages.fetch(value.messageId);
     } catch (error) {
-      const errorCheck = new Error(error);
-      if (errorCheck.message == "DiscordAPIError[10004]: Unknown Guild") {
-        await runDb('DELETE FROM ticket_system_message WHERE guildId = ?', value.guildId);
-      }
-      else if (errorCheck.message == "DiscordAPIError[10003]: Unknown Channel")
-      {
-        await runDb('DELETE FROM ticket_system_message WHERE guildId = ? AND channelId = ?', value.guildId, value.channelId);
-      } 
-      else if (errorCheck.message == "DiscordAPIError[10008]: Unknown Message") {
-        await runDb('DELETE FROM ticket_system_message WHERE guildId = ? AND channelId = ? AND messageId = ?', value.guildId, value.channelId, value.messageId);
+      if(value.initAuthorId == null) {
+        const errorCheck = new Error(error);
+        if (errorCheck.message == "DiscordAPIError[10004]: Unknown Guild") {
+          await runDb('DELETE FROM ticket_system_message WHERE guildId = ?', value.guildId);
+        }
+        else if (errorCheck.message == "DiscordAPIError[10003]: Unknown Channel")
+        {
+          await runDb('DELETE FROM ticket_system_message WHERE guildId = ? AND channelId = ?', value.guildId, value.channelId);
+        } 
+        else if (errorCheck.message == "DiscordAPIError[10008]: Unknown Message") {
+          await runDb('DELETE FROM ticket_system_message WHERE guildId = ? AND channelId = ? AND messageId = ?', value.guildId, value.channelId, value.messageId);
+        }
       }
     }
   })
